pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aiforecastts"
version = "1.0.0"
authors = [{ name = "Tuan Anh Tran", email = "tuanthescientist@gmail.com" }]
description = "Scientific AI-powered Time Series Research Library — HGR, Ensemble, AutoARIMA, Backtesting & AI Agent"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "statsmodels>=0.14",
    "prophet>=1.1",
    "pmdarima>=2.0",
    "xgboost>=2.0",
    "scikit-learn>=1.3",
    "scipy>=1.11",
    "ta>=0.11",
    "joblib>=1.3",
    "google-generativeai>=0.3",
    "matplotlib>=3.7",
]

[project.urls]
Homepage = "https://github.com/tuanthescientist/aiforecastts"
Documentation = "https://github.com/tuanthescientist/aiforecastts#readme"
"Bug Tracker" = "https://github.com/tuanthescientist/aiforecastts/issues"
Changelog = "https://github.com/tuanthescientist/aiforecastts/releases"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black",
    "ruff",
    "build",
    "twine",
]

[tool.setuptools]

[tool.setuptools.packages.find]
where = ["."]
include = ["aiforecastts*"]
